requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
future==0.18.3
idna==3.4
pathlib==1.0.1
Pillow==9.4.0
pycparser==2.21
python-whois==0.8.0
requests==2.28.2
soupsieve==2.3.2.post1
tk==0.1.0
traitlets==5.8.1
ttkbootstrap==1.10.1
urllib3==1.26.14
